chore(wasi): reuse top-level clock types
